Allow overriding branch name with ROCM_DOCS_REMOTE_DETAILS (#26)
There are cases when the automatic branch name detection fails: i.e.
when running on a commit not part of any branch. This could be e.g.
a merge request pipeline that runs on a temporary merge commit.

For these situations allow overriding the automatic detection with
an environment variable
`ROCM_DOCS_REMOTE_DETAILS=<remote-url>,<branch-name>,<allow-edits>`.

Additionally support local development on a branch without a remote,
in this case disable editing, but return the current branch name.