Add automatic API reference support via doxysphinx
The API deliberately doesn't expose the backend of the API to allow
for changing it in the future if needed.
Currently it uses doxysphinx (https://github.com/boschglobal/doxysphinx)

The dependency is behind the `api_reference` library option so that
pure sphinx projects don't need to depend on extra things.

To use the generated API:
1. Specify the option `[api_reference]` for rocm-docs-core in
   requirements.txt.in
2. Install sass (https://sass-lang.com/install) as doxysphinx uses it
   internally.
   This requirement will be removed in the future.
3. calling `run_doxygen` and `enable_api_reference()` from conf.py.
4. Apply the recommended settings [1] for doxysphinx to the doxygen config
   file
5. Add the generated index (index.html corresponding to the doxygen
   output path) file to their TOC

The style of the generated pages is being worked on, as is dark mode
support, these will land in future PRs :).

[1]: https://boschglobal.github.io/doxysphinx/docs/getting_started.html#mandatory-settings