Move theme settings to a custom sphinx theme
As part of making rocm-docs-core a sphinx extension, move its theming
settings to a sphinx theme `rocm_docs_theme`. This makes it easier to
re-use. Also, the theme files no longer need to be copied to
the documentation directory keeping things a little tidier.

This change is transparent to users because `setup()` sets the
sphinx theme to this new bundled theme.