Automatic release (#83)
* Move theme settings to a custom sphinx theme (#41)

As part of making rocm-docs-core a sphinx extension, move its theming
settings to a sphinx theme `rocm_docs_theme`. This makes it easier to
re-use. Also, the theme files no longer need to be copied to
the documentation directory keeping things a little tidier.

This change is transparent to users because `setup()` sets the
sphinx theme to this new bundled theme.

Re-land with fix and requirement for pydata-sphinx-theme 13.3

* Migrate setup.cfg to pyproject.toml

Having just the standard package metadata simplifies things, tools were
sometimes preferring one over the other creating confusion.

Additionally fix the theme folder not being installed

* Add readthedocs build

Add a readthedocs build mostly to verify that the package works, without
needing to create a PR in a dependent repository.

* Make theme settings overrideable

Respect any user defined settings and only provide defaults.
- Move simple defaults to theme.conf, user settings override it by default
- And prefer the users value for the more complex ones

* Move the rest of the visual settings into the theme

Moves the favicon and hiding the sphinx logo into theme, and provide
only a default value (users can now override),

* ci(lint): Add commit message linting via commitizen

The commits of pull requests are now checked with commitizen to conform to conventional commits format.

* ci(release): Add automatic version bump and pypi release

Implement automatic version bumps based on commitizen, every push to main triggers a bump if a new feature/fix is added. The new release is automatically built and published to pypi.

* docs(developer_guide): add short summary of committing rules, commitizen usage