fix(legacy): restore custom theme on readthedocs
Read the docs applies its own theme if the theme is not overriden by the user. Crucially the detection for this happens before `rocm_docs.core` (our main extension) activates meaning that readthedocs doesn't see us setting the theme, and then we (the extension) bail from overriding it as it looks as if a user has set it from conf.py (because that's exactly what read the docs does).