fix(extension): no longer set the html_theme by default in the extension
Because there is no way to tell if the user or read-the-docs has set
the `html_theme`, require the user to explicitly set it and do NOT
provide a default.

This makes local builds more similar to RTD builds,
and makes the theme choice explicit when using rocm-docs-core through
the non-legacy API. The legacy API is unaffected as it will always
overwrite the theme selection made by the user.

BREAKING CHANGE: users of the non-legacy API have to set `html_theme` to `rocm_docs_theme` to maintain the current behaviour.