feat: fetch intersphinx config from a remote file
Intersphinx mappings are loaded from intersphinx_mapping.yaml from the
default branch in this repository using Github's API.
The yaml has the same structure as
the sphinx setting `intersphinx_mapping`, but with arrays instead of
tuples.

If fetching fails the file is also installed as part of rocm-docs-core
and it will be loaded from there. This is to support offline environments.