refactor(projects.py,theme.py): Decouple announcement strings from projects.py
To allow for "flavor" based customization make the interface for
announcements between projects.py and theme.py (the theme and project
sphinx extensions) less coupled:
Instead of passing the announcement text, pass an enumeration that
describes the version status of the version we're building.
The actual text of the announcement is then set in theme.py and can now
depend on the chosen flavor.