fix(core.py): fix setting up the base url for the 404 page
Previously it was erroneously skipped when the relevant config
(`notfound_urls_prefix`) was NOT defined, we should skip when it is
defined to not overwrite the users intention.

Additionally the contained logic didn't account for the sub-project
layout of the urls (i.e. the url for rocm-docs-core is
`/projects/rocm-docs-core/en/latest`). This resulted on 404 pages not
working on branches that don't exists on the main documentation site.