ci(version-bump): Mark status checks as passed on version bump commit
Because of branch protection rules status checks are required before
merging, but they are not triggered by pushes from github actions
(pushes authenticated via `GITHUB_TOKEN`).

Quoting from [Github actions / Using the `GITHUB_TOKEN` in a workflow](https://docs.github.com/en/actions/security-guides/automatic-token-authentication#using-the-github_token-in-a-workflow)

> When you use the repository's `GITHUB_TOKEN` to perform tasks, events
> triggered by the `GITHUB_TOKEN`, with the exception of workflow_dispatch
> and repository_dispatch, will not create a new workflow run.
> ...
> For example, if a workflow run pushes code using the repository's
> `GITHUB_TOKEN`, a new workflow will not run even when the repository
> contains a workflow configured to run when push events occur.

To remedy this and allow the merge back to develop succeed, "skip"
the checks for the bump commit, by marking the required statuse checks
as successful from the github API.
The list of such checks is available from the API, but requires admin
access and cannot be queried using the `GITHUB_TOKEN`.

Therefore to make this work **an action variable must be defined with
the name `REQUIRED_CHECK_CONTEXTS`**. It's value should be the names
of the required checks, one for each line followed by a final newline.

E.g.

```plaintext
Spelling
RestructuredText
Markdown

```

The GitHub docs section [Creating configuration variables for a repository](https://docs.github.com/en/actions/learn-github-actions/variables#creating-configuration-variables-for-a-repository)
explains how to add variables to the repo.