fix(projects): always resolve project references in TOC and templates
Resolve project urls in jinja templates `projects["project_name"]` and
in the TOC (`_toc.yml.in`) regardless of the setting of `external_projects`.

The HTML templates make use of the URL of the ROCm main project, which
were broken with the previous behaviour. Having it available in the TOC
too is more consistent, and there is no extra cost to it, as we already
have this information from `projects.yaml` without downloading the
sphinx inventories.